% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjPicApp.R
\name{adjPicApp}
\alias{adjPicApp}
\title{Title}
\usage{
adjPicApp(pictureIn)
}
\arguments{
\item{pictureIn}{an input photo}
}
\value{
launch an  app that adjusts the picture
}
\description{
Title
}
\examples{
dove = demo_DD()
adjPicApp(dove)
}
