% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustPicture.R
\name{adjustPicture}
\alias{adjustPicture}
\title{Title}
\usage{
adjustPicture(
  pictureIn,
  pixels = 1000,
  imageFlip = FALSE,
  imageFlop = FALSE,
  imageRotate = 0,
  bsh = c(100, 100, 100),
  darkGreyIn = 0.25,
  lightGreyIn = 0.75,
  crop = c(1, 1, 0, 0),
  return = 4,
  imageNegate = FALSE,
  maxPixel = 7000,
  aspect = 0,
  enlarge = 0,
  offCenter = 1,
  offRight = 1,
  backCol1 = 1
)
}
\arguments{
\item{pictureIn}{inphoto}

\item{pixels}{pixels of output photo (long end)}

\item{imageFlip}{Flip image upside down}

\item{imageFlop}{Flop image left to right}

\item{imageRotate}{rotate image}

\item{bsh}{control bsh with 3 element vector}

\item{darkGreyIn}{0 to 1 control dark grey}

\item{lightGreyIn}{0 to 1 control light grey}

\item{crop}{four element vector first two control how much of width and hieght to use second two are how much of right  and top to  chop off}

\item{return}{which of the 6 possiblilities to return (0  to 6) or all off them with 7.}

\item{imageNegate}{take negative of imageNegate}

\item{maxPixel}{overrides pixels to ensure image is not two big}

\item{aspect}{if 0 retains the input if not distorts the image (1 image 2 landscape 0.5 is portrait pixels is the smallest side)}

\item{enlarge}{works with make square}

\item{offCenter}{works with make square}

\item{offRight}{works with make square}

\item{backCol1}{works with make square}
}
\value{
it returns a stack of images
}
\description{
#c(mainPic2,step1,step2,step3,step4,lessContrast,moreContrast)

#return 0 is mainpic2,

step1 is everything lighter than dark grey

step2 is everything darker than light grey and light than dark grey,

step3 is normalizes step 2

step4 is everything darker than light grey

step5 is low contrast

step6 is high contrast
is  3, 4, 5, 6 is moreContrast, 7 is the stack of all
}
\examples{
pictureIn = demo_dd("thescream.jpg")
pictureIn = image_resize(pictureIn,geometry_size_pixels(width=400,heigh=400))
stack = adjustPicture(pictureIn,400,bsh=c(100,100,100),darkGrey = 0.49, lightGrey = 0.50,crop = c(1,1,0,0),return=7,aspect=1)
image_append(stack)
pictureIn = demo_dd("mm.jpg")
adjustPicture(pictureIn,400,bsh=c(100,0,100),darkGrey = 0.4, lightGrey = 0.6,crop = c(1,1,0,0),return=7,aspect=1)
photo = demo_dd("NegWedding.jpg")
adjustPicture(photo, pixels = 800,bsh = c(117,0,101),crop = c(0.47,0.6,0.18,0.16), return = 0, imageNegate = TRUE)
}
