% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_dd.R
\name{demo_dd}
\alias{demo_dd}
\title{demo_dd}
\usage{
demo_dd(path0 = "dove.jpg", pixels = 0)
}
\arguments{
\item{path}{of an image to load}
}
\value{
an image
}
\description{
demo_dd
}
\examples{
demo_dd("dove.jpg")
demo_dd("mm",pixels=400)
}
