% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPixels.R
\name{getPixels}
\alias{getPixels}
\title{Get Pixels}
\usage{
getPixels(pictureIn)
}
\arguments{
\item{pictureIn}{a photograph}
}
\value{
a 2 element numeric vector of the width and hieght of a photo in pixels
}
\description{
Get Pixels
}
\examples{
pictureIn = demo_dd()
cwh= getPixels(pictureIn)
cwh  #cwh is canvas width and height
}
