% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTessalation.R
\name{makeTessalation}
\alias{makeTessalation}
\title{Title}
\usage{
makeTessalation(
  photo,
  rows = 2,
  columns = 2,
  decay = 0,
  center = c(1, 1),
  option = 1,
  print_out = F
)
}
\arguments{
\item{photo}{photograph object to input}

\item{rows}{number of rows in the tessalation}

\item{columns}{number of columns in the tessalation}

\item{decay}{rate at which the tesslation decays}

\item{center}{center of tessalation}

\item{option}{does a variety of options}
}
\value{
a photograph
}
\description{
Title
}
\examples{
#EXAMPLE 1
photo = demo_dd("3wins",pixels=300)
 makeTessalation(photo,rows=3,columns=3,decay = -2, center=c(2,2),option=1)
#EXAMPLE 1
photo = demo_dd("mm",pixels=300)
makeTessalation(photo,rows=1,columns=5,decay = -2, center=c(3,1),option=1)
#EXAMPLE 2
photo = demo_dd("thescream",pixels=300)
makeTessalation(photo,rows=5,columns=3,decay = -2, center=c(2,3),option=1)
#EXAMPLE 3
photo = image_flip(demo_dd("monalisa",pixels=400))
makeTessalation(photo,rows=7,columns=7,decay = -2, center=c(4,4),option=1)
#EXAMPLE 4
photo = image_flip(demo_dd("monalisa",pixels=200))
photo2 = makeTessalation(photo,rows=7,columns=7,decay = 1, center=c(4,4),option=1)
adjustPicture(photo2,pixels=600)
}
