% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transPics.R
\name{transPics}
\alias{transPics}
\title{transPics
Transitions between two photographs gradually}
\usage{
transPics(pic1, pic2, blackNwhite = c(0, 1))
}
\arguments{
\item{pic1}{first photograph}

\item{pic2}{second photograph}

\item{blackNwhite}{a vector that specifies the grey used for the transition.  If c(0,0) returns first image, c(1,0) uses first image on left and second on right}
}
\value{
a photograph
}
\description{
transPics
Transitions between two photographs gradually
}
\examples{
pic1 = demo_dd("mm",pixels=600)
pic2 = demo_dd("che",pixels=600)
transPics(pic1,pic2,blackNwhite = c(1,1))
transPics(pic1,pic2,blackNwhite = c(0,0))
transPics(pic1,pic2,blackNwhite = c(1,0))
transPics(pic1,pic2,blackNwhite = c(0,1))
}
